pyproject.toml
[tool.poetry]
name = "adaptative-rag"
version = "0.1.0"
description = "An intelligent retrieval system with dynamic chunking, hybrid retrieval, and reranking"
authors = ["Vinicius Martins <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
sentence-transformers = "^2.2.2"
faiss-cpu = "^1.7.4"
hydra-core = "^1.3.2"
pandas = "^2.0.0"
numpy = "^1.24.0"
tqdm = "^4.65.0"
scikit-learn = "^1.3.0"
datasets = "^2.14.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
ruff = "^0.0.282"
ipython = "^8.14.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = []

